App modularization

+++ data/drink/src/main/java/com/salazar/cheers/data/drink/mapper/DrinkMapper.kt
@@ -4,7 +4,7 @@ import cheers.drink.v1.Drink
 
 fun Drink.toDrink(): com.salazar.cheers.core.model.Drink {
     return com.salazar.cheers.core.model.Drink(
-        id = id,
+        id = id.toInt(),
         name = name,
         category = category,
         icon = icon,
